// frontend/src/lib/api/models/EarnedBadgeRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EarnedBadgeRead = {
    badgeKey: string;
    family: string;
    title: string;
    description: string;
    icon: string;
    scopeType: string;
    scopeId: string;
    earnedAt: string;
    qualifiedValue?: (number | null);
    threshold?: (number | null);
};