// frontend/src/lib/api/models/LockedBadgeRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LockedBadgeRead = {
    badgeKey: string;
    family: string;
    title: string;
    description: string;
    icon: string;
    scopeType: string;
    scopeId: string;
    current: number;
    target: number;
};